feat(level.ts): now you can be able to see other users level with the level cmd

+++ src/commands/Leveling/level.ts
@@ -3,6 +3,8 @@ import { Message, MessageEmbed } from 'discord.js';
 import { Command } from 'nukejs';
 import { primaryColor } from '../../settings';
 import { getLevelFromXP, getXPFromLevel } from '../../helper/leveling/leveling';
+import { usernameResolver } from '../../helper';
+import { GuildMember } from 'discord.js';
 
 module.exports = class extends Command {
     /**
@@ -13,7 +15,7 @@ module.exports = class extends Command {
             name: 'level',
             category: 'Leveling',
             runIn: ['text'],
-            aliases: ['latency'],
+            aliases: ['levels'],
             botPerms: ['SEND_MESSAGES', 'EMBED_LINKS'],
             description: "Get a user's level.",
             enabled: true,
@@ -28,10 +30,17 @@ module.exports = class extends Command {
      * @param {HozolClient} client
      */
     async run(message: Message, args: string[], client: HozolClient) {
+        if (!message.guild) return;
         await message.delete();
-        const profile = await message.member?.profile();
+        let target = message.member as GuildMember;
+        if (args[0])
+            target = <GuildMember>(
+                await message.guild.members.cache.get((await usernameResolver(message, args.slice(0).join(' '))).id)
+            );
+        if (!target) throw new Error('There was an issue getting this user');
+        const profile = await target.profile();
         const embed = new MessageEmbed()
-            .setAuthor(`${message.author.username}`, `${message.author.displayAvatarURL({ dynamic: true })}`)
+            .setAuthor(`${target.user.username}`, `${target.user.displayAvatarURL({ dynamic: true })}`)
             .setTitle('Leveling Card!')
             .addField(`XP`, `${profile?.XP!}/${getXPFromLevel(getLevelFromXP(profile?.XP!) + 1)}`, true)
             .addField(`Level`, getLevelFromXP(profile?.XP!), true)
@@ -41,7 +50,7 @@ module.exports = class extends Command {
             )
             .setColor(primaryColor)
             .setTimestamp()
-            .setFooter(`User ID: ${message.author.id}`);
+            .setFooter(`User ID: ${target.user.id}`);
         message.channel.send(embed);
     }
 };
